[lipi] navbar fixed
